# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Blog Platform Swagger",
    "version": "1.0.0",
    "description": "This is a sample server blog platform Swagger. You can access the Swagger JSON at [http://localhost:5000/swagger.json](http://localhost:5000/swagger.json)"
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "paths": {
...
    "/categories": {
      "post": {
        "summary": "Add new categories",
        "description": "Add new categories to the system.",
        "tags": [
          "Categories"
        ],
        "requestBody": {
          "description": "Categories data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "categories": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "companyName": {
                    "type": "string"
                  },
                  "productName": {
                    "type": "string"
                  },
                  "modelNumber": {
                    "type": "string"
                  },
                  "size": {
                    "type": "string"
                  },
                  "price": {
                    "type": "number"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Categories added successfully",
            "content": {
              "application/json": {
                "example": {
                  "Status": "Success",
                  "message": "Categories Upload.....!",
                  "data": {
                    "name": "Path",
                    "categories": "electronics",
                    "companyName": "vivo",
                    "productName": "tablets",
                    "modelNumber": "LKI5874WE44D5",
                    "size": "5.5 inch",
                    "price": 22000,
                    "password": 123
                  }
                }
              }
            }
          },
          "409": {
            "description": "Categories already exists",
            "content": {
              "application/json": {
                "example": {
                  "Status": "FAILED",
                  "message": "Categories already exists"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get list of categories",
        "description": "Retrieve a list of categories.",
        "tags": [
          "Categories"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "Bearer <token>"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response containing categories list",
            "content": {
              "application/json": {
                "example": {
                  "Status": "Success",
                  "CategoriesList": {
                    "_id": "64c0e5faaed71535f108ed31",
                    "name": "Path",
                    "categories": "electronics",
                    "companyName": "vivo",
                    "productName": "tablets",
                    "modelNumber": "LKI5874WE44D5",
                    "size": "5.5 inch",
                    "price": 22000,
                    "password": 123
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "example": {
                  "Status": "ERROR",
                  "error": "Error retrieving categories list"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update categories",
        "description": "Update existing categories.",
        "tags": [
          "Categories"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Updated categories data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "categories": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "companyName": {
                    "type": "string"
                  },
                  "productName": {
                    "type": "string"
                  },
                  "modelNumber": {
                    "type": "string"
                  },
                  "size": {
                    "type": "string"
                  },
                  "price": {
                    "type": "number"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Categories updated successfully",
            "content": {
              "application/json": {
                "example": {
                  "Status": "Success",
                  "message": "Categories Updated.....!"
                }
              }
            }
          },
          "404": {
            "description": "Categories not found",
            "content": {
              "application/json": {
                "example": {
                  "Status": "FAILED",
                  "message": "Categories is Exist"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "example": {
                  "Status": "ERROR",
                  "error": "Internal Server Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete categories",
        "description": "Delete an existing categories.",
        "tags": [
          "Categories"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "Bearer <token>"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Categories deleted successfully",
            "content": {
              "application/json": {
                "example": {
                  "Status": "Success",
                  "message": "Categories Record Deleted Successfully"
                }
              }
            }
          },
          "409": {
            "description": "Categories Does Not Exist",
            "content": {
              "application/json": {
                "example": {
                  "Status": "FAILED",
                  "message": "Categories Does Not Exist"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "example": {
                  "Status": "ERROR",
                  "error": "Failed to delete Categories record"
                }
              }
            }
          }
        }
      }
    },
    "/logincategories": {
      "post": {
        "summary": "Login to categories",
        "description": "Login to categories using name and password.",
        "tags": [
          "Categories"
        ],
        "requestBody": {
          "description": "Categories login credentials",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "example": {
                  "Status": "Success",
                  "message": "Login Successfully....!",
                  "data": {
                    "_id": "64c0e5faaed71535f108ed31",
                    "name": "Path",
                    "categories": "electronics",
                    "companyName": "vivo",
                    "productName": "tablets",
                    "modelNumber": "LKI5874WE44D5",
                    "size": "5.5 inch",
                    "price": 22000,
                    "password": 123
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "Status": "FAILED",
                  "message": "Invalid name and password"
                }
              }
            }
          },
          "404": {
            "description": "Categories not found",
            "content": {
              "application/json": {
                "example": {
                  "Status": "FAILED",
                  "message": "User not found"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "example": {
                  "Status": "ERROR",
                  "error": "Internal server error"
                }
              }
            }
          }
        }
      }
    },
    "/postcategories": {
      "get": {
        "summary": "Get paginated categories list",
        "description": "Retrieve a paginated list of categories based on specified filters.",
        "tags": [
          "Categories"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "description": "Page number",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "description": "Number of items per page",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "in": "query",
            "name": "name",
            "description": "Filter categories by name (case-insensitive)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "companyName",
            "description": "Filter categories by companyName (case-insensitive)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "productName",
            "description": "Filter categories by productName (case-insensitive)",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response containing paginated categories list",
            "content": {
              "application/json": {
                "example": {
                  "page": 1,
                  "totalPostes": 1,
                  "totalPages": 1,
                  "postes": {
                    "_id": "64c0e5faaed71535f108ed31",
                    "name": "Path",
                    "categories": "electronics",
                    "companyName": "vivo",
                    "productName": "tablets",
                    "modelNumber": "LKI5874WE44D5",
                    "size": "5.5 inch",
                    "price": 22000,
                    "password": 123
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "example": {
                  "Status": "FAILED",
                  "message": "Internal server error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}